Front wheels are separated from the vehicle
The front wheels are separated from the vehicle (modularization of code).
The car is actually simplified and only has one front wheel, to make the calculations easier.
I actually added the file to git this time...
Improved physics and autopilot -> steering target can be set for the car and it will try to steer
at it's target degree, unless the autopilot is overriden by manual input or autosteering is
disabled.